stitch_prompt_performance_analytics_dashboard/requirements.txt
fastapi
uvicorn[standard]
anthropic
pydantic>=2.0
python-dotenv
tiktoken
aiosqlite
